scraper/requirements.txt
feedparser==6.0.11
anthropic==0.40.0
openai>=1.0.0
google-genai>=1.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pytest==8.3.4
pytest-mock==3.14.0
